add support for hook util, move types
