+++ library/hooks/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@w3ux/hooks-source",
   "license": "GPL-3.0-only",
-  "version": "1.3.1-beta.0",
+  "version": "1.3.1-beta.7",
   "type": "module",
   "scripts": {
     "clear": "rm -rf node_modules dist tsconfig.tsbuildinfo",
@@ -12,7 +12,7 @@
   },
   "devDependencies": {
     "@types/react": "^18",
-    "@w3ux/types": "^0.2.0",
+    "@w3ux/types": "0.2.1-beta.1",
     "gulp": "^5.0.0",
     "gulp-sourcemaps": "^3.0.0",
     "gulp-strip-comments": "^2.6.0",
