+++ library/hooks/src/useTimeleft/defaults.ts
@@ -1,7 +1,7 @@
 /* @license Copyright 2024 w3ux authors & contributors
 SPDX-License-Identifier: GPL-3.0-only */
 
-import { TimeleftDuration } from "./types";
+import type { TimeleftDuration } from "@w3ux/types";
 
 export const defaultDuration: TimeleftDuration = {
   days: 0,
